+++ src/Network/Network/ChangeLog.md
@@ -21,7 +21,14 @@
 
 ## Version 1.5.0
 * Add Threat Intelligence support for Azure Firewall
+* Add Alert action type for Azure Firewall Network and Application Rule Collections
 * Add Application Gateway Firewall Policy top level resource and Custom Rules
+
+* Added support for conditions in RewriteRules in the Application Gateway
+    - New cmdlets added:
+        - New-AzApplicationGatewayRewriteRuleCondition
+    - Cmdlets updated with optional parameter - RuleSequence and Condition
+        - New-AzApplicationGatewayRewriteRule
 
 ## Version 1.4.0
 * Add ResourceId parameter to Get-AzNetworkInterface